Réajustement du contexte d'interaction avec l'espace Tous les espaces devraient être interactifs. Un espace peut être dupliqué, c'est-à-dire que ses formes peuvent être dessinées dans un autre espace ayant son propre repère.
